booking_api-main-main/booking_api-main/stats_streamlit_db/requirements.txt
streamlit==1.37.1
psycopg2-binary==2.9.9
pandas==2.2.2
passlib[bcrypt]==1.7.4